Minor fix for remove post

+++ src/app/components/kehbab-menu/kehbab-menu.component.ts
@@ -11,29 +11,30 @@ import { UserService } from 'src/app/service/user.service';
 export class KehbabMenuComponent implements OnInit {
   @Input() postID: string = '';
   @Input() authorID: string = '';
-  private activeUserRole = 1;
-  private isAuthorized: boolean = false;
+  isAuthorized: boolean = false;
   private session = localStorage.session;
 
   constructor(private router: Router, private crudService: CrudService, private userService: UserService) { }
 
-  ngOnInit( ): void {
+  ngOnInit(): void {
 
     //if user is loggedin, check authority on post
     if (this.session) {
       this.isAuthorized = JSON.parse(localStorage["session"])._id === this.authorID ? true : false;
-      this.userService.getActiveUserRole().subscribe( role => { if (role === 2) this.isAuthorized = true})}
-    
+      this.userService.getActiveUserRole().subscribe(role => { if (role === 2) this.isAuthorized = true })
+    }
+
   }
 
   async removePost() {
-    //delete
-    await this.crudService.DeletePost(this.postID);
+    if (this.isAuthorized){
+      await this.crudService.DeletePost(this.postID);
 
-    //refresh home
-    await this.router.navigateByUrl('/search', { skipLocationChange: true }).then(() => {
-      this.router.navigate(['/home']);
-     }); 
+      //refresh home
+      await this.router.navigateByUrl('/search', { skipLocationChange: true }).then(() => {
+        this.router.navigate(['/home']);
+    });
+  }
   }
 
 }
